Move events folder
